+++ com.unity.cinemachine/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Bugfixes
 - Deoccluder did not always properly reset its state.
+- Deoccluder and Decollider were introducing spurious damping when the FreeLook orbit size changed.
 - Mac only: Extensions dropdown in CinemachineCamera inspector did not work consistently.
 - Regression fix: Confiner2D was not always confining when a camera was newly activated.
 - The RotationComposer no longer damps in response to composition changes from the FreeLookModifier.